encapsulated getting of modules
TestFlightUtil now has methods for getting all of the TestFlight
modules,so its all done in one central place.
Added methods to TestFlightUtil to get the part name and part title.
These are ModuleEngineConfigs aware.
Added new public property to all Modules, TestFlightEnabled which is
used to determine if that module is enabled or not in the TestFlight
system.  This is also MEC aware.
General code cleanup.