* **API Changed**: `baseFailureRate` and `momentaryFailureRate`, as well as momentary failure rate modifiers are all treated internally as doubles now for extra precision with very small failure rates.  All API calls utilizing any of these values now return and expect doubles.
* **NEW Configs**: Added new property to TestFlightCore,
`failureRateModifier` which defaults to 1.  This is a flat, fixed,
modifier applied to the calculated baseFailureRate which can be used to
force smaller values than is possible normally due to 32 bit
restrictions in KSP.
* **NEW**: In order to accommodate very small failure rates, the core
has been reworked to use 64 bit double precision floats internally.
This allows failure rates as low as .00000000000001.  However, KSP
itself will not persist doubles, only floats.  Further more, by their
very nature FloatCurves, which are used to define the failure rate
curves, have to be floats.  This means that while internally the
failure rate can be much lower than before, KSP will still only allow
you to specify failure rate values down to .0000001 (This is one
significant digit lower than previous).  In order to get even smaller
values than that, use the new property on TestFlightCore
`failureRateModifier` to force it even lower.  Setting both the
failureRate and the failureRateModifier to the smallest float value of
.0000001 will result in a calculated failureRate internally of the min
rate, .00000000000001