* **NEW**: `TestFlightReliability_EngineCycle` can now optionally modify accumulated burn time based on the actual thrust output of the engine.  This is defined as a modifier curve using the FloatCurve `thrustModifier`
* **NEW**: `TestFlightReliability_EngineCycle` can now optionally decay
the used burn time on an engine when the engine is turned off.  Can be
defined by using the property `idleDecayRate`.  This value is directly
subtracted from the engine's current burn time each second, so a value
of 1 would be an approximate 1:1 time decay.  This can optionally allow
engines to have extended usage when shut off during coast periods.
Note that this only kicks in if the engine is shut off (not ignited).