Fix survival chance per timestep calculation
TF consistently interprets the 'momentary failure rate' curve as
'instantaneous hazard rate' - the proportional failure rate of systems
still working at the current time - for display purposes (e.g. the
'Reliability' and 'MTBF' displays are calculated as if it is the hazard
rate). However, the conversion to a failure chance per update cycle
wasn't done correctly, and the actual failure rate over time didn't
match the instantaneous MTBF when the failure rate is not constant.

Now, the cycle reliability curve directly specifies the hazard rate h(t)
(which is the same as the 'failure rate' λ(t)), and the chance to fail
over a timestep Δt is h(t)*Δt.